frontend refactor: rename Tabs directory
