+++ frontend/src/components/NewTabs/NewTabs.tsx
