+++ frontend/src/components/PostGrid/PostGrid.tsx
@@ -6,7 +6,7 @@ import { useFetching } from '../../hooks/useFetching'
 import { FeedSection } from '../../types/MockTypes/MockFeedRequest'
 import type { Product } from '../../types/MockTypes/Product'
 import PostCard from '../PostCard/PostCard'
-import NewTabs from '../Tabs/NewTabs'
+import NewTabs from '../NewTabs/NewTabs'
 
 const PostGrid = () => {
 	const [page, setPage] = useState(1)
